feat: move to single review
